requirements.txt
openai==1.12.0
requests==2.31.0
python-dotenv==1.0.1
msal==1.26.0
